refactor: split decorator.py into cohesive decorator package
Split the 1271-line decorator.py monolith into focused submodules under
a new decorator/ package (constants, composition, validation, out,
async_proxies, wrapper, bindings, metadata) while preserving the exact
public and private import surface via a re-export facade in
decorator/__init__.py.

Add characterization tests locking the pre-split contract (import paths,
exported symbols, and per-decorator metadata) so the split cannot
silently change public behavior. Document explicit Tier 1 (supported)
vs Tier 2 (community/BYOD) database support tiers in the README.

Closes #265
